Fixed blue and red paths
