+++ 2024-Competition-Code/src/main/deploy/pathplanner/paths/B1 Left Speaker to Note.path
@@ -3,13 +3,13 @@
   "waypoints": [
     {
       "anchor": {
-        "x": 1.1500679206779225,
-        "y": 6.49
+        "x": 0.75,
+        "y": 6.745441850370149
       },
       "prevControl": null,
       "nextControl": {
-        "x": 2.1500679206779223,
-        "y": 6.49
+        "x": 1.7499999999999996,
+        "y": 6.745441850370149
       },
       "isLocked": false,
       "linkedName": null
@@ -17,11 +17,11 @@
     {
       "anchor": {
         "x": 2.85,
-        "y": 7.25
+        "y": 7.003080254377342
       },
       "prevControl": {
         "x": 1.8044021526373748,
-        "y": 7.25
+        "y": 7.003080254377342
       },
       "nextControl": null,
       "isLocked": false,
