+++ 2024-Competition-Code/src/main/deploy/pathplanner/paths/B1 Note to Speaker.path
@@ -3,13 +3,13 @@
   "waypoints": [
     {
       "anchor": {
-        "x": 2.855,
-        "y": 7.25
+        "x": 2.85,
+        "y": 7.0
       },
       "prevControl": null,
       "nextControl": {
-        "x": 1.8549999999999995,
-        "y": 7.25
+        "x": 1.8499999999999996,
+        "y": 7.0
       },
       "isLocked": false,
       "linkedName": null
