+++ 2024-Competition-Code/src/main/deploy/pathplanner/paths/B3 Middle Speaker to Note.path
@@ -16,11 +16,11 @@
     },
     {
       "anchor": {
-        "x": 2.55,
+        "x": 2.85,
         "y": 4.1
       },
       "prevControl": {
-        "x": 0.5499999999999998,
+        "x": 0.8500000000000001,
         "y": 4.1
       },
       "nextControl": null,
