+++ 2024-Competition-Code/src/main/deploy/pathplanner/paths/B3 Note to Speaker.path
@@ -3,12 +3,12 @@
   "waypoints": [
     {
       "anchor": {
-        "x": 2.55,
+        "x": 2.85,
         "y": 4.1
       },
       "prevControl": null,
       "nextControl": {
-        "x": 0.5499999999999998,
+        "x": 0.8500000000000001,
         "y": 4.1
       },
       "isLocked": false,
