+++ 2024-Competition-Code/src/main/deploy/pathplanner/paths/R1 Left Speaker to Note.path
@@ -4,12 +4,12 @@
     {
       "anchor": {
         "x": 15.753481093265984,
-        "y": 4.333009521940824
+        "y": 4.35
       },
       "prevControl": null,
       "nextControl": {
         "x": 13.993709976809411,
-        "y": 4.304626116838734
+        "y": 4.32161659489791
       },
       "isLocked": false,
       "linkedName": null
@@ -17,11 +17,11 @@
     {
       "anchor": {
         "x": 13.7,
-        "y": 5.55
+        "y": 4.1
       },
       "prevControl": {
-        "x": 14.7999997725216,
-        "y": 5.550707426622337
+        "x": 14.799999999999999,
+        "y": 4.1
       },
       "nextControl": null,
       "isLocked": false,
