+++ 2024-Competition-Code/src/main/deploy/pathplanner/paths/R2 Note to Speaker.path
@@ -16,11 +16,11 @@
     },
     {
       "anchor": {
-        "x": 15.12,
+        "x": 15.1,
         "y": 5.55
       },
       "prevControl": {
-        "x": 14.120000000000001,
+        "x": 14.100000000000001,
         "y": 5.55
       },
       "nextControl": null,
