+++ 2024-Competition-Code/src/main/deploy/pathplanner/paths/R3 Note to Speaker.path
@@ -3,25 +3,25 @@
   "waypoints": [
     {
       "anchor": {
-        "x": 1.4,
-        "y": 5.55
+        "x": 13.7,
+        "y": 7.0
       },
       "prevControl": null,
       "nextControl": {
-        "x": 2.4000000000000004,
-        "y": 5.55
+        "x": 14.799999999999999,
+        "y": 7.0
       },
       "isLocked": false,
       "linkedName": null
     },
     {
       "anchor": {
-        "x": 2.85,
-        "y": 7.0
+        "x": 15.1,
+        "y": 5.55
       },
       "prevControl": {
-        "x": 0.8500000000000001,
-        "y": 7.0
+        "x": 14.100000000000001,
+        "y": 5.55
       },
       "nextControl": null,
       "isLocked": false,
