+++ 2024-Competition-Code/src/main/deploy/pathplanner/paths/R3 Right Speaker to Note.path
@@ -3,24 +3,24 @@
   "waypoints": [
     {
       "anchor": {
-        "x": 1.4,
-        "y": 5.55
+        "x": 15.85,
+        "y": 6.75
       },
       "prevControl": null,
       "nextControl": {
-        "x": 2.4000000000000004,
-        "y": 5.55
+        "x": 14.09146278842819,
+        "y": 6.75
       },
       "isLocked": false,
       "linkedName": null
     },
     {
       "anchor": {
-        "x": 2.85,
+        "x": 13.7,
         "y": 7.0
       },
       "prevControl": {
-        "x": 0.8500000000000001,
+        "x": 14.804903729960223,
         "y": 7.0
       },
       "nextControl": null,
@@ -45,7 +45,7 @@
   "reversed": false,
   "folder": null,
   "previewStartingState": {
-    "rotation": 180.0,
+    "rotation": -59.30027744918538,
     "velocity": 0
   },
   "useDefaultConstraints": true
