+++ 2024-Competition-Code/src/main/java/frc/robot/commands/auto/RedAlliance/Plays/R1_LeftSpeakerNote.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.auto.RedAlliance2.Plays;
+package frc.robot.commands.auto.RedAlliance.Plays;
 
 import com.pathplanner.lib.util.PIDConstants;
 
@@ -17,22 +17,21 @@ import frc.robot.subsystems.swerve.SwerveDrive;
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class R2_RightSpeakerNoteSpeaker extends SequentialCommandGroup {
+public class R1_LeftSpeakerNote extends SequentialCommandGroup {
   SwerveDrive swerve;
   Shooter shooter;
   AutoPath autoPath;
   /** Creates a new B1_StartAmpNoteSpeaker. */
-  public R2_RightSpeakerNoteSpeaker(SwerveDrive swerve, Shooter shooter) {
+  public R1_LeftSpeakerNote(SwerveDrive swerve, Shooter shooter) {
     this.swerve = swerve;
     this.shooter = shooter;
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand());
     addCommands(
-      new Shoot(3500,3500, this.shooter).withTimeout(2),
-      new ParallelCommandGroup(new AutoPath("R2 Right Speaker to Note", this.swerve, new PIDConstants(1.0, 0, 0), new PIDConstants(1.0, 0, 0), true),
-      new IntakeBeamBreak(0.6, this.shooter).withTimeout(4)),
-      new AutoPath("R2 Note to Speaker", this.swerve, new PIDConstants(1.0, 0, 0), new PIDConstants(1.0, 0, 0), false),
-      new Shoot(3500,3500, this.shooter).withTimeout(5)
+      // Tune these high and low rpm
+      new Shoot(0,0, this.shooter).withTimeout(3),
+      new ParallelCommandGroup(new AutoPath("R1 Left Speaker to Note", this.swerve, new PIDConstants(1.0, 0, 0), new PIDConstants(1.0, 0, 0), true),
+      new IntakeBeamBreak(0.6, this.shooter)) //Automatically stops)
     );
   }
 }
