+++ 2024-Competition-Code/src/main/java/frc/robot/commands/auto/RedAlliance/Plays/R_Shoot.java
@@ -2,12 +2,15 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.auto.RedAlliance2.Plays;
+package frc.robot.commands.auto.RedAlliance.Plays;
 
 import com.pathplanner.lib.util.PIDConstants;
 
-import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
+import edu.wpi.first.wpilibj.Timer;
+import edu.wpi.first.wpilibj2.command.Command;
+import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+import frc.robot.RobotContainer;
 import frc.robot.commands.notemechanism.IntakeBeamBreak;
 import frc.robot.commands.notemechanism.Shoot;
 import frc.robot.commands.auto.AutoPath;
@@ -17,22 +20,19 @@ import frc.robot.subsystems.swerve.SwerveDrive;
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class R2_MiddleSpeakerNoteSpeaker extends SequentialCommandGroup {
+public class R_Shoot extends SequentialCommandGroup {
   SwerveDrive swerve;
   Shooter shooter;
   AutoPath autoPath;
   /** Creates a new B1_StartAmpNoteSpeaker. */
-  public R2_MiddleSpeakerNoteSpeaker(SwerveDrive swerve, Shooter shooter) {
+  public R_Shoot(SwerveDrive swerve, Shooter shooter) {
     this.swerve = swerve;
     this.shooter = shooter;
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand());
     addCommands(
-      new Shoot(3500,3500, this.shooter).withTimeout(2),
-      new ParallelCommandGroup(new AutoPath("R2 Middle Speaker to Note", this.swerve, new PIDConstants(1.0, 0, 0), new PIDConstants(1.0, 0, 0), true),
-      new IntakeBeamBreak(0.6, this.shooter).withTimeout(4)),
-      new AutoPath("R2 Note to Speaker", this.swerve, new PIDConstants(1.0, 0, 0), new PIDConstants(1.0, 0, 0), false),
-      new Shoot(3500,3500, this.shooter).withTimeout(5)
+      // *TODO: tune the constants for shooting into the speaker from the left
+      new Shoot(3500, 3500, this.shooter).withTimeout(3)
     );
   }
 }
